Refactor identity values to DefaultIdentity
